+++ Examples/NativeTemplatesRecyclerViewExample/app/src/main/java/com/yandex/nativeads/template/recyclerview/sample/NativeTemplateActivity.java
@@ -22,12 +22,12 @@ import android.widget.Button;
 import com.yandex.mobile.ads.AdRequest;
 import com.yandex.mobile.ads.AdRequestError;
 import com.yandex.mobile.ads.nativeads.NativeAdLoader;
+import com.yandex.mobile.ads.nativeads.NativeAdLoaderConfiguration;
 import com.yandex.mobile.ads.nativeads.NativeAppInstallAd;
 import com.yandex.mobile.ads.nativeads.NativeContentAd;
 
 import java.util.ArrayList;
 import java.util.List;
-
 
 public class NativeTemplateActivity extends AppCompatActivity {
 
@@ -62,11 +62,15 @@ public class NativeTemplateActivity extends AppCompatActivity {
     private void createNativeAdLoader() {
         /*
         * Replace demo R-M-DEMO-native-i with actual Block ID
+        * Please, note, that configured image sizes don't affect demo ads.
         * Following demo Block IDs may be used for testing:
         * app install: R-M-DEMO-native-i
         * content: R-M-DEMO-native-c
         */
-        mNativeAdLoader = new NativeAdLoader(getApplicationContext(), "R-M-DEMO-native-i");
+        final NativeAdLoaderConfiguration adLoaderConfiguration =
+                new NativeAdLoaderConfiguration.Builder("R-M-DEMO-native-c", false)
+                        .setImageSizes(NativeAdLoaderConfiguration.NATIVE_IMAGE_SIZE_LARGE).build();
+        mNativeAdLoader = new NativeAdLoader(getApplicationContext(), adLoaderConfiguration);
         mNativeAdLoader.setOnLoadListener(mNativeAdLoadListener);
     }